# controller-back/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"JsonObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "controller-back",
		"version": "1.0.0",
		"description": "Controller back-end",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/product/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"registrationClient": {
										"type": "string"
									},
									"urlImg": {
										"type": "string"
									},
									"price": {
										"type": "number",
										"format": "double"
									},
									"stock": {
										"type": "number",
										"format": "double"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"registrationClient",
									"urlImg",
									"price",
									"stock",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/product/getAll": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/product/Update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"urlImg": {
										"type": "string"
									},
									"price": {
										"type": "number",
										"format": "double"
									},
									"stock": {
										"type": "number",
										"format": "double"
									},
									"name": {
										"type": "string"
									},
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"urlImg",
									"price",
									"stock",
									"name",
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/product/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/product/getProducts": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "registrationClient",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/product/UpdateStock": {
			"patch": {
				"operationId": "UpdateStock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"stock": {
										"type": "number",
										"format": "double"
									},
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"stock",
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/product/UpdateStockOrder": {
			"patch": {
				"operationId": "UpdateStockOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"stock": {
										"type": "number",
										"format": "double"
									},
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"stock",
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/product/getProduct": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/client/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"themeSelect": {
										"type": "number",
										"format": "double"
									},
									"type": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"registration": {
										"type": "string"
									}
								},
								"required": [
									"themeSelect",
									"type",
									"password",
									"email",
									"registration"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/client/getAll": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/client/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"themeSelect": {
										"type": "number",
										"format": "double"
									},
									"password": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"themeSelect",
									"password",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/client/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/client/getClient": {
			"get": {
				"operationId": "GetClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "newRegistration",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/inventory/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Inventory"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"registrationClient": {
										"type": "string"
									},
									"quant": {
										"type": "number",
										"format": "double"
									},
									"idProduct": {
										"type": "string"
									},
									"data": {
										"type": "string"
									}
								},
								"required": [
									"registrationClient",
									"quant",
									"idProduct",
									"data"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/inventory/getInventories": {
			"get": {
				"operationId": "GetInventories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Inventory"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "registrationClient",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/order/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"products": {
										"items": {
											"properties": {
												"price": {
													"type": "number",
													"format": "double"
												},
												"quant": {
													"type": "number",
													"format": "double"
												},
												"product": {
													"type": "string"
												},
												"idProduct": {
													"type": "string"
												}
											},
											"required": [
												"price",
												"quant",
												"product",
												"idProduct"
											],
											"type": "object"
										},
										"type": "array"
									},
									"priceDesc": {
										"type": "number",
										"format": "double"
									},
									"priceVend": {
										"type": "number",
										"format": "double"
									},
									"strClientVend": {
										"type": "string"
									},
									"registrationClient": {
										"type": "string"
									},
									"numberOrder": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"products",
									"priceDesc",
									"priceVend",
									"strClientVend",
									"registrationClient",
									"numberOrder"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/order/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"products": {
										"items": {
											"properties": {
												"price": {
													"type": "number",
													"format": "double"
												},
												"quant": {
													"type": "number",
													"format": "double"
												},
												"product": {
													"type": "string"
												},
												"idProduct": {
													"type": "string"
												}
											},
											"required": [
												"price",
												"quant",
												"product",
												"idProduct"
											],
											"type": "object"
										},
										"type": "array"
									},
									"priceDesc": {
										"type": "number",
										"format": "double"
									},
									"priceVend": {
										"type": "number",
										"format": "double"
									},
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"products",
									"priceDesc",
									"priceVend",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/order/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/order/getOrders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "registrationClient",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}